Make the integration test_file_path fixture shareable
